+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v0.2.0
 - Add `StateRoutingContext.queryParams`. 
+- Fix bug in logging when current configuration is empty.
 
 ## v0.1.0
 - Initial release version