+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.2.0
 - Add `StateRoutingContext.queryParams`. 
 - Fix bug in logging when current configuration is empty.
+- Categorize documentation
 
 ## v0.1.0
 - Initial release version