+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## v0.2.1
+- Update `tree_state_machine` dependency to 3.0.1
+
 ## v0.2.0
 - Add `StateRoutingContext.queryParams`. 
 - Fix bug in logging when current configuration is empty.